refactor: rename field names
