+++ src/main/java/lotto/model/Lotto.java
@@ -1,13 +1,13 @@
-package lotto;
+package lotto.model;
 
 import java.util.List;
 
 public class Lotto {
-    private final List<Integer> numbers;
+    private final List<Integer> lotto;
 
     public Lotto(List<Integer> numbers) {
         validate(numbers);
-        this.numbers = numbers;
+        lotto = numbers;
     }
 
     private void validate(List<Integer> numbers) {
@@ -16,5 +16,7 @@ public class Lotto {
         }
     }
 
-    // TODO: 추가 기능 구현
+    public List<Integer> getLotto() {
+        return lotto;
+    }
 }
