+++ src/main/java/lotto/model/UserLotto.java
@@ -3,13 +3,13 @@ package lotto.model;
 import java.util.List;
 
 public class UserLotto {
-    private final List<Integer> userNumber;
+    private final List<Integer> userLotto;
 
     public UserLotto(List<Integer> userNumber) {
-        this.userNumber = userNumber.stream().sorted().toList();
+        this.userLotto = userNumber.stream().sorted().toList();
     }
 
-    public List<Integer> getUserNumber() {
-        return userNumber;
+    public List<Integer> getUserLotto() {
+        return userLotto;
     }
 }
